Reformatted to use a Waypoint object for gps locations.  My idea is that usinb objects will allow us to standardize the process, and make it easier to perform additional tasks with our gps locations.
It's a work in progress.

I'm not sure how to handle testing when my function requires serial input.  Maybe I should be gathering the serial input first, then sending it to my function?  That would make sense, except it would make the code much more 'muddy' as I would be moving a lot of logic away from where it's used.  I'll continue to think about it.