Allow ugly map feature using pygmaps
There are some problems using pygmaps.  One, it's generating a new
tab every time a waypoint is added.  Need a way to update the existing
map instead.  Two, it requires an internet connection, which we
won't have.

So, commiting this for now, but I expect we'll want to try going
another direction to map the waypoints.

Also, I began wondering if it shouldn't just log waypoints to a
file, and later map the file all at once as a route?