Fix improper merge with simplecut
Not sure what I did wrong, but clearly I incorrectly merged with the
simplecut branch.  This commit fixes the files, adds a README.md for
both versions of the cutdown, and improves the README.md for the
landing page.

I'll use a demo repo to explore and learn from what went wrong
on the previous merge.